Primarily Install-HealOpsPackage script. Besides, bugfixes and small improvements.
- Bump version & added the Install-HealOpsPackage as exported func.
- Moved global psversion determination and declaration to the main module. So that this code can be shared between functions.
- snippet updated with the latest ver.

Install-HealOps
- null on new temp dir
- improve success output.
- jobtype switch default case update
- update parameter description on jobtype

Install-AvailableUpdate
- make sure that the move-item dest. folder is there when PS v5-

Install-HealOpsPackage
- public func. in HealOps main module used to install additional HealOps packages on a system where HealOps was already deployed

Invoke-HealOps
- small improvements. Spelling mistake
- bugfix on throwing messages when controlling the healops config file
- removed the psversion determination from here to the main module file

+++ Private/UpdateEngine/Install-AvailableUpdate.ps1
@@ -134,11 +134,24 @@ function Install-AvailableUpdate() {
                 }
 
                 if ($installationResult) {
-                    try {
-                        # Move the files from the extraction dir. to the root module dir.
-                        Get-ChildItem -Path $ModuleExtractionPath -Exclude "Artefacts","lib","$ModuleName.zip" -Force -ErrorAction Stop | Move-Item -Destination "$moduleRoot" -Force -ErrorAction Stop
-                    } catch {
-                        $log4netLogger.error("Moving module files to the correct dir failed with > $_")
+                    # Control that our destination exists
+                    if (-not (Test-Path -Path $moduleRoot)) {
+                        try {
+                            New-Item -Path "$psProgramFilesModulesPath" -Name "$ModuleName" -ItemType Directory -Force -ErrorAction Stop | Out-Null
+                        } catch {
+                            $log4netLogger.error("Failed to create the module root folder. The failure was > $_")
+                        }
+                    }
+
+                    if (Test-Path -Path $moduleRoot) {
+                        try {
+                            # Move the files from the extraction dir. to the root module dir.
+                            Get-ChildItem -Path $ModuleExtractionPath -Exclude "Artefacts","lib","$ModuleName.zip" -Force -ErrorAction Stop | Move-Item -Destination "$moduleRoot" -Force -ErrorAction Stop
+                        } catch {
+                            $log4netLogger.error("Moving module files to the correct dir failed with > $_")
+                            $installationResult = $false
+                        }
+                    } else {
                         $installationResult = $false
                     }
                 }
